bill-scanner/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
paddleocr==2.7.3
paddlepaddle==2.6.2
onnxruntime==1.18.0
numpy==1.26.4
faiss-cpu==1.8.0
hdbscan==0.8.38.post2
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.3.0
apscheduler==3.10.4
python-dateutil==2.9.0
pillow==10.3.0
requests==2.32.3
tokenizers==0.19.1